Update README.md
As a junior Data Scientist in Mayor office at the Data Lake town, West Dakota. 314,159 population Growth has occurred in Data Lake as a cause of the remote workers movement to the city. Task given to the office is to learn about the specific needs of the new remote work population, review the data and come back with at least two data-driven policy recommendations of the lives,needs, and habits of remote workers that the office of Mayor will adopt for public policy